Move strategies to their own file
